+++ .gitignore
@@ -40,3 +40,6 @@
 # build directories
 build/
 cmake-*/
+
+# CTest
+Testing/
