+++ tests/unit/general/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_dd_tests(
+        TESTS
+        test_logging
+        test_sample
+)
