--- tests/unit/platf/linux/test_linux_sample.cpp
@@ -1,1 +1,0 @@
-// TODO: remove this file
