--- tests/unit/platf/macos/test_macos_sample.cpp
@@ -1,1 +1,0 @@
-// TODO: remove this file
