+++ tests/unit/windows/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_dd_tests(
+        TESTS
+        test_winapilayer
+
+        LIBRARIES
+        libwindows
+)
